Created Rough NPC Animations
They share an animation controller and so all play animations at the same time. Could be worse but could be a lot better, couldn't be bothered to offset the animations for each of the NPCs. I would need to duplicate the animation controller and add a delay before playing the animations